feat(sqlparser): add materialized view tests and refactor classes
