+++ libs/ob-sql-parser/src/test/java/com/oceanbase/tools/sqlparser/OBMySQLParserTest.java
@@ -35,10 +35,18 @@ import com.oceanbase.tools.sqlparser.statement.Expression;
 import com.oceanbase.tools.sqlparser.statement.Operator;
 import com.oceanbase.tools.sqlparser.statement.Statement;
 import com.oceanbase.tools.sqlparser.statement.common.CharacterType;
+import com.oceanbase.tools.sqlparser.statement.common.ColumnGroupElement;
 import com.oceanbase.tools.sqlparser.statement.common.DataType;
 import com.oceanbase.tools.sqlparser.statement.common.RelationFactor;
+import com.oceanbase.tools.sqlparser.statement.creatematerializedview.CreateMaterializedView;
+import com.oceanbase.tools.sqlparser.statement.creatematerializedview.CreateMaterializedViewOpts;
+import com.oceanbase.tools.sqlparser.statement.creatematerializedview.MaterializedViewRefreshInterval;
+import com.oceanbase.tools.sqlparser.statement.creatematerializedview.MaterializedViewRefreshOnClause;
+import com.oceanbase.tools.sqlparser.statement.creatematerializedview.MaterializedViewRefreshOpts;
 import com.oceanbase.tools.sqlparser.statement.createtable.ColumnDefinition;
 import com.oceanbase.tools.sqlparser.statement.createtable.CreateTable;
+import com.oceanbase.tools.sqlparser.statement.createtable.HashPartition;
+import com.oceanbase.tools.sqlparser.statement.createtable.TableOptions;
 import com.oceanbase.tools.sqlparser.statement.delete.Delete;
 import com.oceanbase.tools.sqlparser.statement.expression.ColumnReference;
 import com.oceanbase.tools.sqlparser.statement.expression.CompoundExpression;
@@ -61,6 +69,53 @@ import com.oceanbase.tools.sqlparser.statement.update.UpdateAssign;
  * @since sqlparser_1.0.0_SNAPSHOT
  */
 public class OBMySQLParserTest {
+
+    @Test
+    public void parse_createMViewStatement_parseSucceed() {
+        SQLParser parser = new OBMySQLParser();
+        Statement actual = parser.parse(new StringReader(
+                "CREATE MATERIALIZED VIEW `zijia`.`test_mv_allsyntax` (PRIMARY KEY (prim)) " +
+                        "DEFAULT CHARSET = gbk ROW_FORMAT = DYNAMIC COMPRESSION = 'zstd_1.3.8' REPLICA_NUM = 1 BLOCK_SIZE = 16384 USE_BLOOM_FILTER = FALSE ENABLE_MACRO_BLOCK_BLOOM_FILTER = FALSE TABLET_SIZE = 134217728 PCTFREE = 10 PARALLEL 5\n"
+                        +
+                        " partition by hash(prim)\n" +
+                        " WITH COLUMN GROUP(all columns, each column) REFRESH COMPLETE ON DEMAND START WITH sysdate() NEXT sysdate() + INTERVAL 1 DAY ENABLE QUERY REWRITE ENABLE ON QUERY COMPUTATION "
+                        +
+                        "AS select `zijia`.`test_mv_base`.`col1` AS `prim`,`zijia`.`test_mv_base`.`col2` AS `col2`,`zijia`.`test_mv_base`.`col3` AS `col3`,`zijia`.`test_mv_base`.`col4` AS `col4` from `zijia`.`test_mv_base`"));
+
+        TableOptions tableOptions = new TableOptions();
+        tableOptions.setParallel(5);
+        tableOptions.setBlockSize(16384);
+        tableOptions.setReplicaNum(1);
+        tableOptions.setUseBloomFilter(false);
+        tableOptions.setTabletSize(134217728);
+        tableOptions.setPctFree(10);
+        tableOptions.setCompression("'zstd_1.3.8'");
+        tableOptions.setRowFormat("DYNAMIC");
+        tableOptions.setCharset("gbk");
+        tableOptions.setEnableMacroBlockBloomFilter(false);
+
+        HashPartition hashPartition = new HashPartition(Collections.singletonList(
+                new ColumnReference(null, null, "prim")), null, null, null);
+
+        List<ColumnGroupElement> columnGroupElements = new ArrayList<>();
+        columnGroupElements.add(new ColumnGroupElement(true, false));
+        columnGroupElements.add(new ColumnGroupElement(false, true));
+
+        MaterializedViewRefreshInterval materializedViewRefreshInterval =
+                new MaterializedViewRefreshInterval("sysdate()", 1L, "DAY");
+        MaterializedViewRefreshOpts materializedViewRefreshOpts = new MaterializedViewRefreshOpts("COMPLETE",
+                materializedViewRefreshInterval, new MaterializedViewRefreshOnClause("DEMAND"));
+        CreateMaterializedViewOpts createMaterializedViewOpts =
+                new CreateMaterializedViewOpts(true, true, materializedViewRefreshOpts);
+
+        CreateMaterializedView createMaterializedView = new CreateMaterializedView();
+        createMaterializedView.setTableOptions(tableOptions);
+        createMaterializedView.setPartition(hashPartition);
+        createMaterializedView.setColumnGroupElements(columnGroupElements);
+        createMaterializedView.setCreateMaterializedViewOpts(createMaterializedViewOpts);
+
+        Assert.assertEquals(createMaterializedView, actual);
+    }
 
     @Test
     public void parse_selectStatement_parseSucceed() {
